Renamed to `FailOn`
