+++ examples/usual/fail_on/example3.rb
@@ -1,11 +1,13 @@
 # frozen_string_literal: true
 
 module Usual
-  module RescueFrom
-    class Example2 < ApplicationService::Base
-      MyException = Class.new(ArgumentError)
+  module FailOn
+    class Example3 < ApplicationService::Base
+      MyException1 = Class.new(ArgumentError)
+      MyException2 = Class.new(ArgumentError)
 
-      fail_on! MyException, with: ->(exception:) { exception.message }
+      fail_on! MyException1, with: ->(exception:) { "#{exception.message} as 1" }
+      fail_on! MyException2, with: ->(exception:) { "#{exception.message} as 2" }
 
       input :invoice_number, type: String
 
@@ -19,7 +21,7 @@ module Usual
       def check_invoice_number!
         return if inputs.invoice_number.start_with?("AA")
 
-        raise MyException, "Invalid invoice number"
+        raise MyException1, "Invalid invoice number"
       end
 
       def assign_invoice_number
